Connection logic finished
Tested with my Arduino Nano on a 1996 Plymouth. Made up a quick and dirty querying thread to send, and added tabs to the GUI page to try and separate data lists from tests and faults.
Data looks good on CCD. SCI needs much love. Arduino is sending data way too quick and making the module mad. But, it connects somewhat reliably on CCD and SCI.