Changed OBD2 connector, refined layout
Decided that it'd be more efficient if I removed all un-used OBD-II pins. Axed about 5 pins and changed the order such that the layout is easier. Managed to get the number of vias down to 68 or so. Separated resistors and capacitors into rows so that I can label them individually.