Widened +12V trace, re-named pins
I had ran a part renaming script before and forgot to change the names of the OBD  pins back to what they're supposed to be.
I'm not sure if it's necessary, but I'm making the trace for the +12v rail significantly wider, since it will carry the most current. At this point I'm unsure if I need to do this to the 3.3 and 5v rails.