Added LED. Fixed soldering issues. 8V IC pins were wrong.
Added an LED for comm indication. Added resistor for this LED. The L78L pinout had VIN and VOUT swapped somehow, fixed this. The MAX17502 has proven very difficult to solder, even with paste. The pad below the chip causes bridging in extreme cases so I fixed that as well.