Scrubbed ISO1050 from board, replaced with MAX13444EASA+
Using the CAN tranceiver was not working with CCD, so I've gone with an RS-485 IC instead. I've tested this IC with CCD and can confirm it works for reading.
Started the process of removing the device names from the board view, instead I will be showing the values for resistors, capacitors, and inductors, for ease of assembly.