Re-write. Fixed SCI. Added ring buffer. Fixed CCD
SCI tested and working as of yesterday. RX and TX on SCI A Engine was tested with multiplexing as it sits on the board schematic. SCI RX has modifications on the schematic that will be committed shortly.
RX buffer was starting to be a problem with the small program memory of the XMEGA 32A4U, so I implemented a ring buffer of 1024 bytes. This can be adjusted at will depending on whether I need the memory later.
CCD RX has been tested and is working with major changes to the schematic. Commits to come. Testing of TX to come as well.