Fixed SCI problems, implemented logging
The queries needed significant re-work for SCI. As it stood, there was tons of extra baud rate changes needed for 1 loop of the query thread. I made a custom function for querying SCI TXItems. Tested on a 1998 Grand Cherokee 5.9L engine. Still lots of work needed to keep the comms from dropping out.